Rework test codes (#334)
