+++ src/engine/core/data/sebeolsik-3sin-1995.yaml
@@ -1,0 +1,97 @@
+Q: $ㅅ$ㅢ
+S-Q: $ㅢ
+W: $ㄹ$ㅑ
+S-W: $ㅑ
+E: $ㅂ$ㅕ
+S-E: $ㅕ
+R: $ㄷ$ㅐ
+S-R: $ㅐ
+T: $ㅌ$ㅓ
+S-T: $ㅓ
+A: $ㅇ$ㅒ
+S-A: $ㅒ
+S: $ㄴ$ㅖ
+S-S: $ㅖ
+D: $ㅎ$ㅣ
+S-D: $ㅣ
+F: $ㅈ$ㅏ
+S-F: $ㅏ
+G: $ㅍ$ㅡ
+S-G: $ㅡ
+Z: $ㅁ$ㅠ
+S-Z: $ㅠ
+X: $ㄱ$ㅛ
+S-X: $ㅛ
+C: $ㅊ$ㅔ
+S-C: $ㅔ
+V: $ㅋ$ㅗ
+S-V: $ㅗ
+B: $ㅆ$ㅜ
+S-B: $ㅜ
+
+Y: ㄹ
+S-Y: ㄹ
+U: ㄷ
+S-U: ㄷ
+I: ㅁㅜ
+S-I: ㅜ
+O: ㅊㅜ
+S-O: ㅜ
+P: ㅍㅗ
+S-P: ㅗ
+H: ㄴ
+S-H: ㄴ
+J: ㅇ
+S-J: ;
+K: ㄱ
+S-K: "'"
+L: ㅈ
+S-L: ㅈ
+SemiColon: ㅂ
+S-SemiColon: ':'
+Quote: ㅌ
+S-Quote: '"'
+N: ㅅ
+S-N: ㅅ
+M: ㅎ
+S-M: /
+Slash: ㅋ
+S-Slash: '?'
+
+Grave: '`'
+S-Grave: '~'
+1: 1
+2: 2
+3: 3
+4: 4
+5: 5
+6: 6
+7: 7
+8: 8
+9: 9
+0: 0
+
+S-1: '!'
+S-2: '@'
+S-3: '#'
+S-4: '$'
+S-5: '%'
+S-6: '^'
+S-7: '&'
+S-8: '*'
+S-9: '('
+S-0: ')'
+Minus: '-'
+S-Minus: _
+Equal: '='
+S-Equal: +
+Backslash: \
+S-Backslash: '|'
+OpenBracket: '['
+S-OpenBracket: '{'
+CloseBracket: ']'
+S-CloseBracket: '}'
+Comma: ','
+S-Comma: '<'
+Period: '.'
+S-Period: '>'
