+++ src/engine/core/tests/sebeolsik_3_90.rs
@@ -1,47 +1,7 @@
-use kime_engine_core::{Config, InputEngine, InputResult, Key, KeyCode::*, RawConfig};
+#[macro_use]
+mod shared;
 
-#[track_caller]
-fn test_input(keys: &[(Key, &str, &str)]) {
-    let config = Config::from_raw_config(
-        RawConfig {
-            layout: "sebeolsik-390".into(),
-            ..Default::default()
-        },
-        None,
-    );
-
-    let mut engine = InputEngine::new(false);
-
-    engine.set_hangul_enable(true);
-
-    for (key, preedit, commit) in keys.iter().copied() {
-        eprintln!("Key: {:?}", key);
-
-        let ret = engine.press_key(key, &config);
-
-        eprintln!("Ret: {:?}", ret);
-
-        if ret.contains(InputResult::HAS_PREEDIT) {
-            assert_eq!(preedit, engine.preedit_str());
-        } else {
-            assert!(preedit.is_empty());
-        }
-
-        if !ret.contains(InputResult::CONSUMED) {
-            assert_eq!(commit, format!("{}PASS", engine.commit_str()));
-        } else if ret.intersects(InputResult::NEED_RESET | InputResult::NEED_FLUSH) {
-            assert_eq!(commit, engine.commit_str());
-        } else {
-            assert!(commit.is_empty());
-        }
-
-        if ret.contains(InputResult::NEED_RESET) {
-            engine.reset();
-        } else if ret.contains(InputResult::NEED_FLUSH) {
-            engine.flush();
-        }
-    }
-}
+define_layout_test!("sebeolsik-3-90");
 
 #[test]
 fn hello() {
